refactor: rename o1js.js to snarky.js across multiple files for better code clarity and understanding
